chore: move AddLanguagePackage and AddVocab to own dir
