+++ src/App.jsx
@@ -3,12 +3,11 @@ import { Route, HashRouter, Redirect } from "react-router-dom";
 import "./Components/sections.scss";
 import { useSelector } from "react-redux";
 import Image from "./images/vocascan-logo.png";
-import { makeStyles } from "@material-ui/core/styles";
 import Box from "@material-ui/core/Box";
 import "./index.scss";
 
 import Nav from "./Components/Nav/Nav.jsx";
-import AddVocab from "./screens/AddVocab.jsx";
+import AddVocab from "./screens/AddVocab/AddVocab.jsx";
 import SelectionScreen from "./screens/SelectionScreen/SelectionScreen.jsx";
 import Login from "./screens/Login/Login.jsx";
 import Register from "./screens/Register/Register.jsx";
@@ -17,7 +16,6 @@ import TopNav from "./Components/Nav/TopNav.jsx";
 
 function App() {
   const isLoggedIn = useSelector((state) => state.login.isLoggedIn);
-  const classes = useStyles();
 
   if (!isLoggedIn) {
     return (
@@ -33,7 +31,7 @@ function App() {
   } else {
     return (
       <HashRouter>
-        <Box className={classes.root}>
+        <Box className="root">
           <Nav />
           <TopNav />
           <FirstStartup />
@@ -43,17 +41,5 @@ function App() {
     );
   }
 }
-
-const useStyles = makeStyles({
-  root: {
-    display: "grid",
-    width: "100%",
-    height: "100vh",
-    gridTemplateColumns: "200px auto",
-    gridTemplateRows: "40px auto",
-    margin: 0,
-    padding: 0,
-  },
-});
 
 export default App;
