+++ src/index.scss
@@ -7,5 +7,14 @@
 }
 
 .root {
-  display: flex;
+  display: grid;
+  width: 100%;
+  height: 100vh;
+  grid-template-columns: 200px auto;
+  grid-template-rows: 40px auto;
+  grid-template-areas:
+    "sidenav topnav"
+    "sidenav main";
+  margin: 0;
+  padding: 0;
 }
