+++ src/screens/AddVocab/AddVocab.scss
@@ -1,0 +1,53 @@
+@import "../../colors";
+
+.add-vocab-form {
+  display: flex;
+  flex-direction: column;
+  grid-area: main;
+  padding: 50px;
+  max-width: 600px;
+  justify-content: center;
+  margin: 0 auto;
+
+  .heading {
+    grid-column: 2;
+    grid-row: 2;
+  }
+
+  .dropdowns {
+    height: 50px;
+    display: flex;
+    justify-content: space-between;
+    z-index: 4;
+    margin: 20px -20px;
+
+    .form-control {
+      margin: 0 20px;
+      min-width: 200px;
+      width: 100%;
+    }
+  }
+
+  .input-fields {
+    grid-column: 2;
+    grid-row: 4;
+    display: flex;
+    flex-direction: column;
+
+    .submit-btn {
+      width: 150px;
+      height: 35px;
+      grid-column: 2;
+      grid-row: 5;
+      background: $color-primary;
+      box-shadow: -1px 3px 5px $color-alternative;
+      margin: 0 auto;
+    }
+  }
+
+  .form-submit {
+    display: flex;
+    justify-content: center;
+    margin-top: 40px;
+  }
+}
